Refactor names of services
